requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
websockets==11.0.3
pydantic==2.4.2

# LLM integration
llama-cpp-python==0.2.11
transformers==4.34.1
torch==2.1.0

# Web scraping
playwright==1.39.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1

# Storage
sqlalchemy==2.0.22
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1